Fix experience eating for scaling bingos
In minecraft renaming item adds +1 level requirement for all anvil operations.

As workbench uses anvil recipe simulator to display outcome per tick, the option to set name just eats all player levels the longer workbench is opened. 

Setting level cost to -1 is not a pretty solution, but it is the one that works 100% as resulting cost will always be 0.